Split sidebar into multiple components. Simplify layout.
