+++ package.json
@@ -17,6 +17,7 @@
     "axios": "^0.21.1",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
+    "cytoscape": "^3.29.2",
     "express": "^4.17.1",
     "gremlin": "^3.4.4",
     "lodash": "^4.17.15",
