+++ public/index.html
@@ -15,7 +15,7 @@
 
       .mynetwork {
         width: 100%;
-        height:100vh;
+        height:98vh;
         border: 1px solid lightgray;
       }
